Add example of streamification of a discrete API
In this particular case, we 'streamify' the Telegram Bot API with an Iterator
that keeps an internal buffer of elements retrieved from the API.

On each `next` call, pop the first element of the buffer.
When there are no more elements in the buffer, request the API the next
Update batch.

If no Updates are retrieved, or if the connection is dropped, try again
with a backoff, either until we get a new Update or disconnect.

A simple use case is then presented in rx-tg-test, where we filter all
messages that are "f" and then subscribe to it.